requirements.txt
python-dotenv
pylint